More iteration: getting the mod-morph into the mix.
I am not quite sure why the tap half of the hold-tap is the one that
gets called by the macro. The behavior that I observe is:
 - short press, no mods: small 'q'
 - short press, shift: big 'Q'
 - long press, no mods: lowercase 'qz' (the z is intentional for debugging)
 - long press, shift: just Q uppercase 'Qu'
So mostly working as intended, except that the interaction between
macro and hold-tap is confusing. Maybe the macro is activating the
behavior as though it were a tap, but that doesn't explain why neither
half of the hold-tap activates when I short-press the macro.