Fiddling with hold-taps for Qu and shift-enter.
New shift-enter hold-tap puts tappable S-enter on the left big thumb
key in the symbol layer, moving space to a hold behavior on the same
key. This way I can do shift-enter for multiline stuff in chat
programs and still have a space key in the SYM layer, just in case.

For the Qu combo, I want to try a simpler hold-tap again, to see if it
helps with fast typing.