Take away parallelism.
Because so many of my targets are phony, it is hard to declare good
dependency rules. Without good dependency rules, Make does not know
when to wait for a prerequisite before proceeding, and can end up
doing things like starting to build one half of a split board while
`make clean` is still working, or trying to start a transfer or deploy
before both halves of a split board are done. Rather than try to get
the dependencies just right, I am instead simply removing the
parallelism.