Putting F and V back to plain keypresses.
Added a combo to toggle F for prolonged item pickup.
For whatever reason, the slight delay from tap-dances bothers me,
even when I know to expect it. I can rebind the H (horizontal flip)
command to F in Factorio and it will work fine, so I do not need H
on the left hand side of the split.